Play with computer
Add PlayerX and Player0 to game.
Add toggles on UI for each player to toggle between human and computer.

After each move, the game will check if the next player is computer.
If the next player is the computer, then it choose the next move:
1. place a winning move (if it can win)
2. block the opponent if he can win next move (if the opponent can win)
3. random move

Limitations in this commit:
The computer move is added instantly, without delay.
The computer cannot start the game.